Breakup sidebar components
