+++ packages/desktop-client/src/components/sidebar/Accounts.tsx
@@ -1,0 +1,154 @@
+import React, { useState, useMemo } from 'react';
+
+import Add from '../../icons/v1/Add';
+import View from '../common/View';
+
+import { Account } from './Account';
+import { SecondaryItem } from './SecondaryItem';
+
+const fontWeight = 600;
+
+export function Accounts({
+  accounts,
+  failedAccounts,
+  updatedAccounts,
+  getAccountPath,
+  allAccountsPath,
+  budgetedAccountPath,
+  offBudgetAccountPath,
+  getBalanceQuery,
+  getAllAccountBalance,
+  getOnBudgetBalance,
+  getOffBudgetBalance,
+  showClosedAccounts,
+  onAddAccount,
+  onToggleClosedAccounts,
+  onReorder,
+}) {
+  let [isDragging, setIsDragging] = useState(false);
+  let offbudgetAccounts = useMemo(
+    () =>
+      accounts.filter(
+        account => account.closed === 0 && account.offbudget === 1,
+      ),
+    [accounts],
+  );
+  let budgetedAccounts = useMemo(
+    () =>
+      accounts.filter(
+        account => account.closed === 0 && account.offbudget === 0,
+      ),
+    [accounts],
+  );
+  let closedAccounts = useMemo(
+    () => accounts.filter(account => account.closed === 1),
+    [accounts],
+  );
+
+  function onDragChange(drag) {
+    setIsDragging(drag.state === 'start');
+  }
+
+  let makeDropPadding = (i, length) => {
+    if (i === 0) {
+      return {
+        paddingTop: isDragging ? 15 : 0,
+        marginTop: isDragging ? -15 : 0,
+      };
+    }
+    return null;
+  };
+
+  return (
+    <View>
+      <Account
+        name="All accounts"
+        to={allAccountsPath}
+        query={getAllAccountBalance()}
+        style={{ fontWeight, marginTop: 15 }}
+      />
+
+      {budgetedAccounts.length > 0 && (
+        <Account
+          name="For budget"
+          to={budgetedAccountPath}
+          query={getOnBudgetBalance()}
+          style={{ fontWeight, marginTop: 13 }}
+        />
+      )}
+
+      {budgetedAccounts.map((account, i) => (
+        <Account
+          key={account.id}
+          name={account.name}
+          account={account}
+          connected={!!account.bankId}
+          failed={failedAccounts && failedAccounts.has(account.id)}
+          updated={updatedAccounts && updatedAccounts.includes(account.id)}
+          to={getAccountPath(account)}
+          query={getBalanceQuery(account)}
+          onDragChange={onDragChange}
+          onDrop={onReorder}
+          outerStyle={makeDropPadding(i, budgetedAccounts.length)}
+        />
+      ))}
+
+      {offbudgetAccounts.length > 0 && (
+        <Account
+          name="Off budget"
+          to={offBudgetAccountPath}
+          query={getOffBudgetBalance()}
+          style={{ fontWeight, marginTop: 13 }}
+        />
+      )}
+
+      {offbudgetAccounts.map((account, i) => (
+        <Account
+          key={account.id}
+          name={account.name}
+          account={account}
+          connected={!!account.bankId}
+          failed={failedAccounts && failedAccounts.has(account.id)}
+          updated={updatedAccounts && updatedAccounts.includes(account.id)}
+          to={getAccountPath(account)}
+          query={getBalanceQuery(account)}
+          onDragChange={onDragChange}
+          onDrop={onReorder}
+          outerStyle={makeDropPadding(i, offbudgetAccounts.length)}
+        />
+      ))}
+
+      {closedAccounts.length > 0 && (
+        <SecondaryItem
+          style={{ marginTop: 15 }}
+          title={'Closed accounts' + (showClosedAccounts ? '' : '...')}
+          onClick={onToggleClosedAccounts}
+          bold
+        />
+      )}
+
+      {showClosedAccounts &&
+        closedAccounts.map((account, i) => (
+          <Account
+            key={account.id}
+            name={account.name}
+            account={account}
+            to={getAccountPath(account)}
+            query={getBalanceQuery(account)}
+            onDragChange={onDragChange}
+            onDrop={onReorder}
+          />
+        ))}
+
+      <SecondaryItem
+        style={{
+          marginTop: 15,
+          marginBottom: 9,
+        }}
+        onClick={onAddAccount}
+        Icon={Add}
+        title="Add account"
+      />
+    </View>
+  );
+}
